refactor(argocd): extract git write-back engine out of models
Move the git write-back retry engine (UpdateGitImageTag,
generateOverrideFileContent, extractManagedImages, runGitUpdateWithRetry,
backoff, cache invalidation, the ErrDeploymentSuperseded sentinel, and
the two managed-image annotation constants) from internal/models/argo.go
into a new internal/argocd/git_writeback.go.

This removes the models -> internal/updater dependency: the domain model
no longer imports an infrastructure package to carry an orchestration
engine that hung off *Application only because it was a method. The
engine lands in argocd, which already depends on both models and updater,
so the leaf updater package stays a leaf.

Behavior-preserving: functions moved verbatim (receiver converted to a
parameter, models. qualifiers added); the generateOverrideFileContent
receiver is dropped since it never used it. Engine tests move to
internal/argocd (git_writeback_test.go, git_writeback_backoff_test.go,
and the renamed integration tests); the domain-method tests stay in
models. Test coverage is unchanged (22 test functions preserved).
